+++ flutter-hms-analytics/ios/Classes/AnalyticsPlugin.m
@@ -1,5 +1,5 @@
 /*
-    Copyright 2020. Huawei Technologies Co., Ltd. All rights reserved.
+    Copyright 2020-2021. Huawei Technologies Co., Ltd. All rights reserved.
 
     Licensed under the Apache License, Version 2.0 (the "License")
     you may not use this file except in compliance with the License.
@@ -26,6 +26,6 @@
 #endif
 @implementation AnalyticsPlugin
 + (void)registerWithRegistrar:(NSObject<FlutterPluginRegistrar>*)registrar {
-  [AnalyticsPluginSwift registerWithRegistrar:registrar];
+  [HMSAnalyticsMethodCallHandler registerWithRegistrar:registrar];
 }
 @end