+++ flutter-hms-analytics/ios/Classes/Logger.swift
@@ -1,5 +1,5 @@
-﻿/*
- Copyright 2020. Huawei Technologies Co., Ltd. All rights reserved.
+/*
+ Copyright 2020-2021. Huawei Technologies Co., Ltd. All rights reserved.
  
  Licensed under the Apache License, Version 2.0 (the "License");
  you may not use this file except in compliance with the License.
@@ -19,20 +19,20 @@ import SystemConfiguration
 import CoreTelephony
 import Foundation
 
-let Log = Logger(version: "5.0.4.300", service: "FlutterAnalytics")
+let Log = Logger(version: "5.1.0.300", service: "FlutterAnalytics")
 
 /// Custom Logger Class to listen api log events from RN Side.
 final class Logger {
-  
+
   /// Info Types to declare success and failure scenarios.
-  public enum InfoType{
+  public enum InfoType {
     case call, success, fail
   }
-  
+
   private var prefix: String = GenericFunctions.className(Logger.self)
   private var version: String?
   private var service: String?
-  
+
   /// Version and Service names may be added to give more detail in panel.
   /// - Parameters:
   ///   - version: Refers to version number of the framework.
@@ -41,7 +41,7 @@ final class Logger {
     self.version = version
     self.service = service
   }
-  
+
   /// This function can be used during synchronized events. It gets the method name, calculates start end end times and reports to console as single events.
   /// - Parameters:
   ///   - file: Refers to file name that the function is in.
@@ -51,14 +51,14 @@ final class Logger {
   ///   - message: Refers to message that can be added in the panel.
   ///   - block: Refers to function that will be used.
   /// - Returns: Void
-  func debug(file: String = #file, line: Int = #line, name: String = #function, _ message: String, block:()->()) {
-      showInPanel(file:file, line:line, name:name, message: message, type: .call)
+  func debug(file: String = #file, line: Int = #line, name: String = #function, _ message: String, block:() -> Void) {
+      showInPanel(file: file, line: line, name: name, message: message, type: .call)
       block()
-      showInPanel(file:file, line:line, name:name, message: message, type: .success)
+      showInPanel(file: file, line: line, name: name, message: message, type: .success)
   }
-  
+
   // MARK: - Private Functions
-  
+
   /// Shows infos in panel
   private func showInPanel(callTime: Double? = nil,
                            file: String = #file,
@@ -77,21 +77,21 @@ final class Logger {
       printStatus(name, false)
     }
   }
-  
+
   /// Prints call messages among with the function names.
-  private func printCall(_ message: String){
+  private func printCall(_ message: String) {
     print("call \(message)")
   }
-  
+
   /// Print status messages.
-  private func printStatus(_ name: String, _ status: Bool){
+  private func printStatus(_ name: String, _ status: Bool) {
     let statusMsg = status ? "success" : "failure"
     print("\(name) \(statusMsg)")
   }
-  
+
   /// Returns file name of the function is in.
   private func getFileName(_ filePath: String) -> String {
-    let parser = filePath.split(separator:"/")
+    let parser = filePath.split(separator: "/")
     if let fileName = String(parser.last ?? "").split(separator: ".").first {
       return String(fileName)
     }
@@ -121,7 +121,7 @@ extension UIDevice {
 }
 
 extension Bundle {
-  
+
   public var packageName: String {
     if let result = Bundle.main.bundleIdentifier {
       return result
@@ -129,7 +129,7 @@ extension Bundle {
       return "⚠️"
     }
   }
-  
+
   public var appVersionShort: String {
     if let result = infoDictionary?["CFBundleShortVersionString"] as? String {
       return result