+++ flutter-hms-analytics/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 5.1.0+300 
+
+* Added the setRestrictionEnabled, isRestrictionEnabled and deleteUserProfile methods.
+* Added getReportPolicyThreshold. This function is specifically used by Android Platforms.
+* Update for the demo project.
+* Added the HAUserProfileType.
+* Update HAEventType and HAParamType.
+
 ## 5.0.3+300 
 
 * Added the enableLogger and disableLogger methods.